Fixed bug of Aqi Converter
Fixed bug of aqi converter
 : aqicn uses airnow’s table of AQI and formula
move convert function to aqi.converter.js
#2085
